third_party/blink/web_tests/http/tests/devtools/profiler/cpu-profiler-profiling-expected.txt

Tests that CPU profiling works. https://bugs.webkit.org/show_bug.cgi?id=52634

Profiler was enabled.

Running: testProfiling
found pageFunction cpu-profiler-profiling.js:16
found (anonymous) cpu-profiler-profiling.js:0

Profiler was disabled.
